Hack the kaleidoscope image generation
Hacks the kaleidoscope image generation in the case
of some values needed to recalculate it are set to 'NaN'.
Why this happens is still a mystery, for now we enjoy the
hack.
Also, improved the 'requestAnimationFrame' thingy so that
it does not try to repaint forever and ever. My browser
will be very happy about it... and the battery of your
mobile device too!
	modified:   src/scripts/kaleidoscope.js